chore(GameTable): GameTable 패키지 이동
